mod MatchItemResultChip move to Badge in designSystem module
